+++ packages/devtools_app/lib/src/shared/config_specific/server/_server_stub.dart
@@ -99,6 +99,9 @@ Future<ExtensionEnabledState> extensionEnabledState({
   );
 }
 
+Future<List<String>> requestAndroidBuildVariants(String path) async =>
+    const <String>[];
+
 void logWarning() {
   throw Exception(unsupportedMessage);
 }